Remove SchedulerV2, inline all into Scheduler
