+++ Source/EasyNetQ.Scheduler.Mongo.Core/EasyNetQ.Scheduler.Mongo.Core.csproj
@@ -49,12 +49,9 @@
       <Link>Properties\Version.cs</Link>
     </Compile>
     <Compile Include="ISchedulerService.cs" />
-    <Compile Include="ScheduleV2.cs" />
-    <Compile Include="ScheduleV2Repository.cs" />
     <Compile Include="ScheduleRepository.cs" />
     <Compile Include="IScheduleRepositoryConfiguration.cs" />
     <Compile Include="SchedulerService.cs" />
-    <Compile Include="SchedulerV2Service.cs" />
     <Compile Include="ISchedulerServiceConfiguration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Schedule.cs" />
