+++ Source/EasyNetQ.Scheduler.Mongo.Core/Schedule.cs
@@ -9,12 +9,21 @@ namespace EasyNetQ.Scheduler.Mongo.Core
 
         public DateTime WakeTime { get; set; }
 
+        [Obsolete]
         public string BindingKey { get; set; }
 
         [BsonIgnoreIfNull]
         public string CancellationKey { get; set; }
 
+        public string Exchange { get; set; }
+
+        public string ExchangeType { get; set; }
+
+        public string RoutingKey { get; set; }
+
         public byte[] InnerMessage { get; set; }
+
+        public MessageProperties BasicProperties { get; set; }
 
         public ScheduleState State { get; set; }
