+++ Source/EasyNetQ.Scheduler.Mongo.Core/SchedulerService.cs
@@ -66,13 +66,13 @@ namespace EasyNetQ.Scheduler.Mongo.Core
                     if (schedule == null)
                         return;
                     log.DebugWrite(string.Format("Publishing Scheduled Message with Routing Key: '{0}'", schedule.BindingKey));
-                    var exchange = bus.Advanced.ExchangeDeclare(schedule.BindingKey, ExchangeType.Topic);
+                    var exchange = bus.Advanced.ExchangeDeclare(schedule.Exchange ?? schedule.BindingKey, schedule.ExchangeType ?? ExchangeType.Topic);
                     bus.Advanced.Publish(
                         exchange,
-                        schedule.BindingKey,
+                        schedule.RoutingKey ?? schedule.BindingKey,
                         false,
                         false,
-                        new MessageProperties {Type = schedule.BindingKey},
+                        schedule.BasicProperties ?? new MessageProperties {Type = schedule.BindingKey},
                         schedule.InnerMessage);
                     scheduleRepository.MarkAsPublished(schedule.Id);
                     ++published;
@@ -94,14 +94,18 @@ namespace EasyNetQ.Scheduler.Mongo.Core
         {
             log.DebugWrite("Got Schedule Message");
             scheduleRepository.Store(new Schedule
-                {
-                    Id = Guid.NewGuid(),
-                    CancellationKey = message.CancellationKey,
-                    BindingKey = message.BindingKey,
-                    InnerMessage = message.InnerMessage,
-                    State = ScheduleState.Pending,
-                    WakeTime = message.WakeTime
-                });
+            {
+                Id = Guid.NewGuid(),
+                CancellationKey = message.CancellationKey,
+                BindingKey = message.BindingKey,
+                InnerMessage = message.InnerMessage,
+                State = ScheduleState.Pending,
+                WakeTime = message.WakeTime,
+                Exchange = message.Exchange,
+                ExchangeType = message.ExchangeType,
+                RoutingKey = message.RoutingKey,
+                BasicProperties = message.MessageProperties
+            });
         }
     }
 }
