+++ Source/EasyNetQ/SystemMessages/ScheduleMe.cs
@@ -6,8 +6,14 @@ namespace EasyNetQ.SystemMessages
     public class ScheduleMe
     {
         public DateTime WakeTime { get; set; }
-        public string BindingKey { get; set; }
         public string CancellationKey { get; set; }
+        public string Exchange { get; set; }
+        public string ExchangeType { get; set; }
+        public string RoutingKey { get; set; }
         public byte[] InnerMessage { get; set; }
+        public MessageProperties MessageProperties { get; set; }
+
+        [Obsolete]
+        public string BindingKey { get; set; }
     }
 }
